added InMemoryClinicalNote used by FNE plus eliminated extra methods in shortcut class hierarchy
